+++ Engine/src/main/java/com/ing/engine/commands/webservice/Webservice.java
@@ -1,21 +1,14 @@
 package com.ing.engine.commands.webservice;
 
 import com.ing.datalib.settings.DriverProperties;
-import com.ing.engine.commands.browser.General;
-import com.ing.engine.constants.FilePath;
 import com.ing.engine.core.CommandControl;
 import com.ing.engine.core.Control;
-import com.ing.engine.execution.exception.ActionException;
 import com.ing.ingenious.api.status.Status;
 import com.ing.ingenious.api.annotation.Action;
 import com.ing.ingenious.api.types.InputType;
 import com.ing.ingenious.api.types.ObjectType;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.net.InetSocketAddress;
-import java.net.ProxySelector;
-import java.net.URI;
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Collection;
@@ -25,8 +18,6 @@ import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathFactory;
 import com.jayway.jsonpath.*;
-import java.io.FileInputStream;
-import java.io.FileWriter;
 import java.util.ArrayList;
 import java.util.List;
 import org.w3c.dom.Document;
@@ -34,30 +25,14 @@ import org.w3c.dom.NodeList;
 import org.w3c.dom.Node;
 import org.xml.sax.InputSource;
 import java.io.StringReader;
-import java.net.URLEncoder;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpRequest.BodyPublisher;
-import java.net.http.HttpResponse;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.security.KeyStore;
-import java.security.SecureRandom;
-import java.security.cert.X509Certificate;
-import java.time.Duration;
-import java.time.Instant;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import javax.net.ssl.KeyManager;
-import javax.net.ssl.KeyManagerFactory;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.TrustManagerFactory;
-import javax.net.ssl.X509TrustManager;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
+import javax.xml.xpath.XPathFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.xpath.XPathExpressionException;
 import org.json.simple.JSONArray;
@@ -66,25 +41,16 @@ import org.json.simple.parser.JSONParser;
 import org.w3c.dom.DOMException;
 import org.xml.sax.SAXException;
 
-public class Webservice extends General {
+public class Webservice extends GeneralWebservice {
 
     public Webservice(CommandControl cc) {
         super(cc);
     }
 
-    public enum RequestMethod {
-        POST,
-        PUT,
-        PATCH,
-        GET,
-        DELETE,
-        DELETEWITHPAYLOAD
-    }
-
     @Action(object = ObjectType.WEBSERVICE, desc = "PUT Rest Request ", input = InputType.YES, condition = InputType.OPTIONAL)
     public void putRestRequest() {
         try {
-            createhttpRequest(RequestMethod.PUT);
+            createHttpRequest(RequestMethod.PUT);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -95,7 +61,7 @@ public class Webservice extends General {
     @Action(object = ObjectType.WEBSERVICE, desc = "POST Rest Request ", input = InputType.YES, condition = InputType.OPTIONAL)
     public void postRestRequest() {
         try {
-            createhttpRequest(RequestMethod.POST);
+            createHttpRequest(RequestMethod.POST);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -106,7 +72,7 @@ public class Webservice extends General {
     @Action(object = ObjectType.WEBSERVICE, desc = "POST SOAP Request ", input = InputType.YES, condition = InputType.OPTIONAL)
     public void postSoapRequest() {
         try {
-            createhttpRequest(RequestMethod.POST);
+            createHttpRequest(RequestMethod.POST);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -117,7 +83,7 @@ public class Webservice extends General {
     @Action(object = ObjectType.WEBSERVICE, desc = "PATCH Rest Request ", input = InputType.YES, condition = InputType.OPTIONAL)
     public void patchRestRequest() {
         try {
-            createhttpRequest(RequestMethod.PATCH);
+            createHttpRequest(RequestMethod.PATCH);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -128,7 +94,7 @@ public class Webservice extends General {
     @Action(object = ObjectType.WEBSERVICE, desc = "GET Rest Request ", input = InputType.NO, condition = InputType.OPTIONAL)
     public void getRestRequest() {
         try {
-            createhttpRequest(RequestMethod.GET);
+            createHttpRequest(RequestMethod.GET);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -139,7 +105,7 @@ public class Webservice extends General {
     @Action(object = ObjectType.WEBSERVICE, desc = "DELETE Rest Request ", input = InputType.NO)
     public void deleteRestRequest() {
         try {
-            createhttpRequest(RequestMethod.DELETE);
+            createHttpRequest(RequestMethod.DELETE);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -150,7 +116,7 @@ public class Webservice extends General {
     @Action(object = ObjectType.WEBSERVICE, desc = "DELETE with Payload ", input = InputType.YES)
     public void deleteWithPayload() {
         try {
-            createhttpRequest(RequestMethod.DELETEWITHPAYLOAD);
+            createHttpRequest(RequestMethod.DELETEWITHPAYLOAD);
         } catch (Exception e) {
             Report.updateTestLog(Action,
                     "An unexpected error occurred while executing the request : " + "\n" + e.getMessage(),
@@ -453,28 +419,6 @@ public class Webservice extends General {
         }
     }
 
-    private void returnResponseDetails() throws IOException, InterruptedException {
-
-        initiateClientBuilder();
-        sslCertificateVerification();
-        handleProxy();
-
-        /**
-         * *** need to add timeout,version******
-         */
-        httpClient.put(key, httpClientBuilder.get(key).build());
-        httpRequest.put(key, httpRequestBuilder.get(key).build());
-        response.put(key, httpClient.get(key).send(httpRequest.get(key), HttpResponse.BodyHandlers.ofString()));
-
-        responsebodies.put(key, (String) response.get(key).body());
-
-        after.put(key, Instant.now());
-        savePayload("response", (String) response.get(key).body());
-
-        responsecodes.put(key, Integer.toString(response.get(key).statusCode()));
-
-    }
-
     @Action(object = ObjectType.WEBSERVICE, desc = "Assert JSON Element Count ", input = InputType.YES, condition = InputType.YES)
     public void assertJSONelementCount() {
 
@@ -682,38 +626,6 @@ public class Webservice extends General {
 
     }
 
-    private boolean isformUrlencoded() {
-        if (headers.containsKey(key)) {
-            ArrayList<String> headerlist = headers.get(key);
-            if (headerlist.size() > 0) {
-                for (String header : headerlist) {
-                    if (header.split("=")[1].contains("x-www-form-urlencoded")) {
-                        return true;
-                    }
-                };
-            }
-        }
-        return false;
-    }
-
-    private String urlencodedParams() {
-        Map<String, String> parameters = new HashMap<>();
-        String urlParamString = "";
-        try {
-            ArrayList<String> params = urlParams.get(key);
-            for (String param : params) {
-                parameters.put(param.split("=", 2)[0], param.split("=", 2)[1]);
-            }
-            urlParamString = parameters.entrySet()
-                    .stream()
-                    .map(e -> e.getKey() + "=" + URLEncoder.encode(e.getValue(), StandardCharsets.UTF_8))
-                    .collect(Collectors.joining("&"));
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-        return urlParamString;
-    }
-
     @Action(object = ObjectType.WEBSERVICE, desc = "Close the connection ", input = InputType.NO)
     public void closeConnection() {
         try {
@@ -729,358 +641,6 @@ public class Webservice extends General {
             Logger.getLogger(this.getClass().getName()).log(Level.OFF, null, ex);
             Report.updateTestLog(Action, "Error closing connection :" + "\n" + ex.getMessage(), Status.DEBUG);
         }
-    }
-    
-    private ProxySelector getProxyDetails() {
-        if (Control.getCurrentProject().getProjectSettings().getDriverSettings().useProxy()) {
-            String proxyhost = Control.getCurrentProject().getProjectSettings().getDriverSettings()
-                    .getProxyHost().replaceFirst("^(http://|https://)", "");
-            String proxyport = Control.getCurrentProject().getProjectSettings().getDriverSettings()
-                    .getProxyPort();
-            ProxySelector proxySelector = ProxySelector.of(new InetSocketAddress(proxyhost, Integer.parseInt(proxyport)));
-            return proxySelector;
-        } else {
-            return null;
-        }
-    }
-
-    private String getHttpAgentDetails() {
-        if (Control.getCurrentProject().getProjectSettings().getUserDefinedSettings().stringPropertyNames()
-                .contains("http.agent")) {
-            if (!getUserDefinedData("http.agent").isEmpty()) {
-                httpagents.put(key, getUserDefinedData("http.agent"));
-                return httpagents.get(key);
-            } else {
-                return null;
-            }
-        } else {
-            return null;
-        }
-    }
-
-    private String handlePayloadorEndpoint(String data) throws FileNotFoundException {
-        String payloadstring = data;
-        payloadstring = handleDataSheetVariables(payloadstring);
-        payloadstring = handleuserDefinedVariables(payloadstring);
-        System.out.println("Payload :" + payloadstring);
-        return payloadstring;
-    }
-
-    private String handleDataSheetVariables(String payloadstring) {
-        List<String> sheetlist = Control.getCurrentProject().getTestData().getTestDataFor(Control.exe.runEnv())
-                .getTestDataNames();
-        for (int sheet = 0; sheet < sheetlist.size(); sheet++) {
-            if (payloadstring.contains("{" + sheetlist.get(sheet) + ":")) {
-                com.ing.datalib.testdata.model.TestDataModel tdModel = Control.getCurrentProject()
-                        .getTestData().getTestDataByName(sheetlist.get(sheet));
-                List<String> columns = tdModel.getColumns();
-                for (int col = 0; col < columns.size(); col++) {
-                    if (payloadstring.contains("{" + sheetlist.get(sheet) + ":" + columns.get(col) + "}")) {
-                        payloadstring = payloadstring.replace("{" + sheetlist.get(sheet) + ":" + columns.get(col) + "}",
-                                userData.getData(sheetlist.get(sheet), columns.get(col)));
-                    }
-                }
-            }
-        }
-        return payloadstring;
-    }
-
-    private String handleuserDefinedVariables(String payloadstring) {
-        Collection<Object> valuelist = Control.getCurrentProject().getProjectSettings().getUserDefinedSettings()
-                .values();
-        for (Object prop : valuelist) {
-            if (payloadstring.contains("{" + prop + "}")) {
-                payloadstring = payloadstring.replace("{" + prop + "}", prop.toString());
-            }
-        }
-        return payloadstring;
-    }
-
-    private void OpenURLconnection() {
-        try {
-            httpRequestBuilder.put(key, HttpRequest.newBuilder());
-            URI uri = URI.create(endPoints.get(key));
-            httpRequestBuilder.put(key, httpRequestBuilder.get(key).uri(uri));
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    private void setheaders() {
-        try {
-            if (headers.containsKey(key)) {
-                ArrayList<String> headerlist = headers.get(key);
-                System.out.println(headerlist);
-                if (headerlist.size() > 0) {
-                    headerlist.forEach((header) -> {
-                        httpRequestBuilder.put(key, httpRequestBuilder.get(key).setHeader(header.substring(0, header.indexOf("=")), header.substring(header.indexOf("=") + 1, header.length())));
-                    });
-                }
-            }
-
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    private void httpAgentCheck() {
-        try {
-            if (getHttpAgentDetails() != null) {
-                System.setProperty("http.agent", getHttpAgentDetails());
-            }
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    private boolean isMultiPart() {
-        if (headers.containsKey(key)) {
-            ArrayList<String> headerlist = headers.get(key);
-            if (headerlist.size() > 0) {
-                for (String header : headerlist) {
-                    if (header.split("=")[1].contains("multipart")) {
-                        return true;
-                    }
-                };
-            }
-        }
-        return false;
-    }
-
-    private void setRequestMethod(String method, String payload) throws IOException {
-        BodyPublisher payloadBody = null;
-        if (isformUrlencoded()) {
-            payload = urlencodedParams();
-        }
-        if (isMultiPart()) {
-            Path filePath = Path.of(getVar("%filePath%"));
-            filePath = Path.of(Control.getCurrentProject().getLocation() + "/" + filePath);
-            String mimeType = Files.probeContentType(filePath);
-            System.out.println("Path of the file === " + filePath);
-            String boundary = "Boundary-" + System.currentTimeMillis();
-            String fileName = filePath.getFileName().toString();
-
-            /* String body = "--" + boundary.getBytes(StandardCharsets.UTF_8)+ "\r\n"
-                    + "Content-Disposition: form-data; name=\"file\"; filename=\"" + fileName + "\"\r\n"
-                    + "Content-Type: " + mimeType // Set Content-Type to text/csv
-                    + Files.readString(filePath, StandardCharsets.UTF_8) + "\r\n"
-                    + ("--" + boundary + "--\r\n").getBytes(StandardCharsets.UTF_8);*/
-            var byteArrays = new ArrayList<byte[]>();
-            byteArrays.add(("--" + boundary + "\r\n").getBytes(StandardCharsets.UTF_8));
-            byteArrays.add(("Content-Disposition: form-data; name=\"file\"; filename=\"" + fileName + "\"\r\n").getBytes(StandardCharsets.UTF_8));
-            byteArrays.add(("Content-Type: " + mimeType + "\r\n\r\n").getBytes(StandardCharsets.UTF_8));
-            byteArrays.add(Files.readAllBytes(filePath));
-            byteArrays.add(("\r\n--" + boundary + "--\r\n").getBytes(StandardCharsets.UTF_8));
-
-            payloadBody = HttpRequest.BodyPublishers.ofByteArrays(byteArrays);
-            httpRequestBuilder.put(key, httpRequestBuilder.get(key).setHeader("Content-Type", "multipart/form-data; boundary=" + boundary));
-        } else {
-            payloadBody = HttpRequest.BodyPublishers.ofString(payload);
-        }
-        try {
-            switch (method) {
-                case "POST": {
-                    httpRequestBuilder.put(key, httpRequestBuilder.get(key).POST(payloadBody));
-                    savePayload("request", payload);
-                    break;
-                }
-                case "PUT": {
-                    httpRequestBuilder.put(key, httpRequestBuilder.get(key).PUT(payloadBody));
-                    savePayload("request", payload);
-                    break;
-                }
-                case "PATCH": {
-                    httpRequestBuilder.put(key, httpRequestBuilder.get(key).method("PATCH", payloadBody));
-                    savePayload("request", payload);
-                    break;
-                }
-                case "GET": {
-                    httpRequestBuilder.put(key, httpRequestBuilder.get(key).GET());
-                    break;
-                }
-                case "DELETE": {
-                    httpRequestBuilder.put(key, httpRequestBuilder.get(key).DELETE());
-                    break;
-                }
-                case "DELETEWITHPAYLOAD": {
-                    httpRequestBuilder.put(key, httpRequestBuilder.get(key).method("DELETE", payloadBody));
-                    savePayload("request", payload);
-                    break;
-                }
-
-            }
-            headers.remove(key);
-            urlParams.remove(key);
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    private void setRequestMethod(RequestMethod requestmethod) throws FileNotFoundException, IOException {
-        if (requestmethod.toString().equals("PUT") || requestmethod.toString().equals("POST") || requestmethod.toString().equals("PATCH") || requestmethod.toString().equals("DELETEWITHPAYLOAD")) {
-
-            setRequestMethod(requestmethod.toString(), handlePayloadorEndpoint(Data));
-        } else {
-
-            setRequestMethod(requestmethod.toString(), "");
-        }
-    }
-
-    private void createhttpRequest(RequestMethod requestmethod) throws InterruptedException, Exception {
-        try {
-            setheaders();
-            setRequestMethod(requestmethod);
-            before.put(key, Instant.now());
-
-            returnResponseDetails();
-            duration.put(key, Duration.between(before.get(key), after.get(key)).toMillis());
-            Report.updateTestLog(Action, "Response received in : [" + duration.get(key) + "ms] with Status code  : " + responsecodes.get(key), Status.COMPLETE);
-
-            if (headers.containsKey(key)) {
-                if (!headers.get(key).isEmpty()) {
-                    headers.get(key).clear();
-                }
-            }
-
-        } catch (IOException ex) {
-            int responseCode = 0;
-            Matcher exMsgStatusCodeMatcher = Pattern.compile("^Server returned HTTP response code: (\\d+)")
-                    .matcher(ex.getMessage());
-
-            if (exMsgStatusCodeMatcher.find()) {
-                responseCode = Integer.parseInt(exMsgStatusCodeMatcher.group(1));
-            } else if (ex.getClass().getSimpleName().equals("FileNotFoundException")) {
-                System.out.println("\n =====================================\n" + " Returned [FileNotFoundException]" + "\n =====================================\n");
-                responseCode = 404;
-
-            } else {
-                System.out.println(
-                        "Exception (" + ex.getClass().getSimpleName() + ") doesn't contain status code: " + ex);
-            }
-            if (responseCode == 0) {
-                System.out.println("\n =====================================\n" + "Response Code does not exist in Exception" + "\n =====================================\n");
-            } else {
-                responsecodes.put(key, Integer.toString(responseCode));
-            }
-
-            if (responseCode == 400 || responseCode == 401 || responseCode == 402 || responseCode == 403
-                    || responseCode == 404) {
-                Report.updateTestLog(Action,
-                        "Error in executing [" + requestmethod.toString() + "] request : " + "\n" + ex.getMessage(),
-                        Status.DONE);
-
-            } else {
-                Report.updateTestLog(Action,
-                        "Error in executing " + requestmethod.toString() + " request : " + "\n" + ex.getMessage(),
-                        Status.DEBUG);
-            }
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-            throw new ActionException(e);
-        }
-    }
-
-    private void savePayload(String reqOrRes, String data) {
-        String payloadFileName = "";
-        String path = "";
-        if (reqOrRes.equals("request")) {
-            payloadFileName = Report.getWebserviceRequestFileName();
-        } else if (reqOrRes.equals("response")) {
-            payloadFileName = Report.getWebserviceResponseFileName();
-
-        }
-        try {
-            if (!payloadFileName.isBlank()) {
-                path = FilePath.getCurrentResultsPath() + File.separator + "webservice";
-                File file = new File(path);
-                file.mkdirs();
-                //FileManager.mkdir(path);
-                File location = new File(FilePath.getCurrentResultsPath() + payloadFileName);
-                if (location.createNewFile()) {
-                    FileWriter writer = new FileWriter(location);
-                    writer.write(data);
-                    writer.close();
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void handleProxy() {
-        try {
-            if (getProxyDetails() != null) {
-                System.out.println("\nRequest opened with following proxy details :\n" + getProxyDetails().toString() + "\n");
-                httpClientBuilder.put(key, httpClientBuilder.get(key).proxy(getProxyDetails()));
-            }
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    private void initiateClientBuilder() {
-        try {
-            httpClientBuilder.put(key, HttpClient.newBuilder().version(HttpClient.Version.HTTP_1_1));
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    TrustManager[] trustAllCerts = new TrustManager[]{new X509TrustManager() {
-        @Override
-        public X509Certificate[] getAcceptedIssuers() {
-            return null;
-        }
-
-        @Override
-        public void checkClientTrusted(X509Certificate[] certs, String authType) {
-        }
-
-        @Override
-        public void checkServerTrusted(X509Certificate[] certs, String authType) {
-        }
-    }};
-
-    private KeyManager[] loadKeyStore() {
-        String keystorePath = Control.getCurrentProject().getProjectSettings().getDriverSettings().getProperty("keyStorePath");
-        String keystorePassword = Control.getCurrentProject().getProjectSettings().getDriverSettings().getProperty("keyStorePassword");
-        KeyStore keyStore;
-        KeyManagerFactory kmf = null;
-        try {
-            keyStore = KeyStore.getInstance("JKS");
-            keyStore.load(new FileInputStream(keystorePath), keystorePassword.toCharArray());
-            kmf = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
-            kmf.init(keyStore, keystorePassword.toCharArray());
-            TrustManagerFactory tmf = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
-            tmf.init(keyStore);
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-        return kmf.getKeyManagers();
-    }
-
-    private void sslCertificateVerification() {
-        try {
-            if (!isSSLCertificateVerification()) {
-                SSLContext sc = SSLContext.getInstance("TLS");
-                if (isSelfSigned()) {
-                    sc.init(loadKeyStore(), trustAllCerts, new SecureRandom());
-                } else {
-                    sc.init(null, trustAllCerts, new SecureRandom());
-                }
-                httpClientBuilder.put(key, httpClientBuilder.get(key)).sslContext(sc);
-            }
-        } catch (Exception ex) {
-            Logger.getLogger(this.getClass().getName()).log(Level.OFF, ex.getMessage(), ex);
-        }
-    }
-
-    private Boolean isSSLCertificateVerification() {
-        return Control.getCurrentProject().getProjectSettings().getDriverSettings().sslCertificateVerification();
-    }
-
-    private Boolean isSelfSigned() {
-        return Control.getCurrentProject().getProjectSettings().getDriverSettings().selfSigned();
     }
 
 }
