+++ ingenious-api/src/main/java/com/ing/ingenious/api/contract/WebservicePluginApi.java
@@ -39,5 +39,17 @@ public interface WebservicePluginApi extends CommandPluginApi  {
      * @return the HTTP user agent
      */
     String HttpAgent();
+    
+    /**
+     * Creates and executes an HTTP request with the specified request method.
+     * This method handles the complete lifecycle of an HTTP request including
+     * setting headers, configuring the request method, executing the request,
+     * and capturing response details.
+     * 
+     * @param requestMethod the HTTP request method (POST, PUT, PATCH, GET, DELETE, DELETEWITHPAYLOAD)
+     * @throws InterruptedException if the request is interrupted
+     * @throws Exception if an error occurs during request execution
+     */
+    void createHttpRequest(Object requestMethod) throws InterruptedException, Exception;
 
 }
